# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DogDex API - Pokemon GO for dogs
servers:
  - url: /api
    description: Base API path
...
paths:
  /badge-image:
    post:
      operationId: startBadgeImage
      tags: [badgeImage]
      summary: Start generating a badge share image
      description: >
        Queues an AI badge image for the authenticated user and returns
        immediately. Generation continues server-side, so the client can leave
        and return later.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartBadgeImageRequest"
      responses:
        "200":
          description: Generation state after the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeImageStatus"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /badge-image/{badgeId}:
    get:
      operationId: getBadgeImageStatus
      tags: [badgeImage]
      summary: Get badge share image status
      parameters:
        - name: badgeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current generation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeImageStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    StartBadgeImageRequest:
      type: object
      properties:
        badgeId:
          type: string
        breeds:
          type: array
          description: Most recently collected breed names, newest first
          items:
            type: string
        totalCollected:
          type: integer
          minimum: 1
          description: Total number of dogs collected by the user
        regenerate:
          type: boolean
          description: Discard any existing image and generate a new one
      required:
        - badgeId
        - breeds
        - totalCollected
    BadgeImageStatus:
      type: object
      properties:
        status:
          type: string
          description: none = never requested, pending = generating
          enum: [none, pending, ready, failed]
        objectPath:
          type: string
          nullable: true
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message